spin_trigger
Module cli

Modules

help

Structs

FactorsConfig

Configuration options that are common to all triggers.

FactorsTriggerCommand

A command that runs a TriggerExecutor.

InitialKvSetterHook

An [ExecutorHooks] that sets initial key-value pairs in the default store.

KeyValueDefaultStoreSummaryHook

An [ExecutorHooks] that prints information about the default KV store.

LaunchMetadata

Contains information about the trigger flags (and potentially in future configuration) that a consumer (such as spin up) can query using --launch-metadata-only.

NoCliArgs

An empty implementation of clap::Args to be used as TriggerExecutor::RunConfig for executors that do not need additional CLI args.

SqlStatementExecutorHook

ExecutorHook for executing sqlite statements.

SqliteDefaultStoreSummaryHook

An [ExecutorHooks] that prints information about the default KV store.

StdioLoggingExecutorHooks

Implements TriggerHooks, writing logs to a log file and (optionally) stderr

TriggerAppBuilder

A builder for a TriggerApp.

Enums

UserProvidedPath

A user provided option which be either be provided, default, or explicitly none.

Constants

APP_LOG_DIR

DISABLE_WASMTIME_CACHE

FOLLOW_LOG_OPT

RUNTIME_CONFIG_FILE

SPIN_LOCAL_APP_DIR

SPIN_LOCKED_URL

SPIN_WORKING_DIR

WASMTIME_CACHE_FILE

Traits

RuntimeFactorsBuilder

A builder for runtime factors.