spin_trigger::cli
Struct FactorsTriggerCommand

pub struct FactorsTriggerCommand<T: Trigger<B::Factors>, B: RuntimeFactorsBuilder> {
    pub log: Option<PathBuf>,
    pub disable_cache: bool,
    pub cache: Option<PathBuf>,
    pub disable_pooling: bool,
    pub follow_components: Vec<String>,
    pub silence_component_logs: bool,
    pub runtime_config_file: Option<PathBuf>,
    pub state_dir: Option<String>,
    pub trigger_args: T::CliArgs,
    pub builder_args: B::CliArgs,
    pub help_args_only: bool,
    pub launch_metadata_only: bool,
}

A command that runs a TriggerExecutor.

Fields
log: Option<PathBuf>

Log directory for the stdout and stderr of components. Setting to the empty string disables logging to disk.

disable_cache: bool

Disable Wasmtime cache.

cache: Option<PathBuf>

Wasmtime cache configuration file.

disable_pooling: bool

Disable Wasmtime’s pooling instance allocator.

follow_components: Vec<String>

Print output to stdout/stderr only for given component(s)

silence_component_logs: bool

Silence all component output to stdout/stderr

runtime_config_file: Option<PathBuf>

Configuration file for config providers and wasmtime config.

state_dir: Option<String>

Set the application state directory path. This is used in the default locations for logs, key value stores, etc.

For local apps, this defaults to .spin/ relative to the spin.toml file. For remote apps, this has no default (unset). Passing an empty value forces the value to be unset.

trigger_args: T::CliArgsbuilder_args: B::CliArgshelp_args_only: boollaunch_metadata_only: bool
Implementations

impl<T: Trigger<B::Factors>, B: RuntimeFactorsBuilder> FactorsTriggerCommand<T, B>

pub async fn run(self) -> Result<()>

Create a new TriggerExecutorBuilder from this TriggerExecutorCommand.